third_party/WebKit/LayoutTests/inspector-protocol/dom/dom-relayout-boundary-expected.txt

Relayout boundary for div#outer is: html
Relayout boundary for div#boundary is: div#boundary
Relayout boundary for div#inner is: div#boundary
Relayout boundary for div#hidden is: No renderer for node, perhaps orphan or hidden node
